CRAY! : Input formatting day 5! 🙄
For day 5, I really wonder if the aim was for the actual
issue was for the elves to receive any help.

I mean, they really don't deserve it! Their input was poor
and terrible.

Shamefully, I spent about four hours over 2 days trying to
get the input in a sane way for me to use.

It's still terrible but it works!

Signed-off-by: Fon E. Noel NFEBE <fenn25.fn@gmail.com>