Day 2 Debug point
At this point, I was wondering what was wrong everything seemed
fined based on my defination of the problem.

I decided to debug my outputs againts : https://dev.to/doctorlai/avent-of-code-day-02-rock-paper-scissors-27d4

After comparing the outputs for the 2nd and 3rd lines it turns out I
forgot and important part of the problem.

Even though rock awards just a point (1) it still beats scissors which offers 3 points.

So translating each input and using it's weight would always result in a win for playing
scissors over rock!

APOLOGIES: The first time I read about ROCK PAPER SCISSORS was on this issue, so,
I really missed the rules for the game.

Well, well, HOPEFULLY that was it.

Signed-off-by: Fon E. Noel NFEBE <fenn25.fn@gmail.com>